+++ src/main/java/roomescape/reservation/application/dto/response/ReservationDetailFindResponse.java
@@ -2,8 +2,8 @@ package roomescape.reservation.application.dto.response;
 
 import roomescape.reservation.domain.ReservationStatus;
 import roomescape.reservation.application.port.out.projection.ReservationDetailProjection;
-import roomescape.reservationtime.dto.response.TimeInformation;
-import roomescape.theme.dto.response.ThemeFindResponse;
+import roomescape.reservationtime.application.dto.response.TimeInformation;
+import roomescape.theme.application.dto.response.ThemeFindResponse;
 import roomescape.waiting.application.port.out.projection.WaitingDetailProjection;
 
 import java.time.LocalDate;
