--- src/main/java/roomescape/reservationtime/ReservationTime.java
@@ -1,6 +1,0 @@
-package roomescape.reservationtime;
-
-import java.time.LocalTime;
-
-public record ReservationTime(Long id, LocalTime startAt) {
-}
