+++ src/main/java/roomescape/reservationtime/application/ReservationTimeService.java
@@ -10,14 +10,14 @@ import org.springframework.stereotype.Service;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
 import roomescape.reservation.application.port.out.ReservationRepository;
-import roomescape.reservationtime.ReservationTime;
-import roomescape.reservationtime.dto.request.ReservationTimeSaveRequest;
-import roomescape.reservationtime.dto.response.AvailableTimeFindResponse;
-import roomescape.reservationtime.dto.response.ReservationTimeFindResponse;
-import roomescape.reservationtime.dto.response.ReservationTimeSaveResponse;
-import roomescape.reservationtime.dto.response.TimeInformation;
-import roomescape.reservationtime.dto.response.TimeSlotStatus;
-import roomescape.reservationtime.infrastructure.ReservationTimeRepository;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.reservationtime.application.dto.request.ReservationTimeSaveRequest;
+import roomescape.reservationtime.application.dto.response.AvailableTimeFindResponse;
+import roomescape.reservationtime.application.dto.response.ReservationTimeFindResponse;
+import roomescape.reservationtime.application.dto.response.ReservationTimeSaveResponse;
+import roomescape.reservationtime.application.dto.response.TimeInformation;
+import roomescape.reservationtime.application.dto.response.TimeSlotStatus;
+import roomescape.reservationtime.application.port.out.ReservationTimeRepository;
 import roomescape.slot.application.SlotUsageValidator;
 import roomescape.waiting.application.port.out.WaitingRepository;
