--- src/main/java/roomescape/reservationtime/dto/request/ReservationTimeSaveRequest.java
@@ -1,11 +1,0 @@
-package roomescape.reservationtime.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.validation.constraints.NotNull;
-
-import java.time.LocalTime;
-
-public record ReservationTimeSaveRequest(
-        @JsonFormat(pattern = "HH:mm") @NotNull LocalTime startAt
-) {
-}
