--- src/main/java/roomescape/reservationtime/dto/response/AvailableTimeFindResponse.java
@@ -1,7 +1,0 @@
-package roomescape.reservationtime.dto.response;
-
-public record AvailableTimeFindResponse(
-        TimeInformation timeInformation,
-        TimeSlotStatus status
-) {
-}
