--- src/main/java/roomescape/reservationtime/dto/response/ReservationTimeFindResponse.java
@@ -1,20 +1,0 @@
-package roomescape.reservationtime.dto.response;
-
-import roomescape.reservationtime.ReservationTime;
-
-import java.time.LocalTime;
-import java.util.List;
-
-public record ReservationTimeFindResponse(
-        Long id,
-        LocalTime startAt
-) {
-    public static List<ReservationTimeFindResponse> from(List<ReservationTime> reservationTimes) {
-        return reservationTimes.stream()
-                .map(reservationTime -> new ReservationTimeFindResponse(
-                        reservationTime.id(),
-                        reservationTime.startAt()
-                ))
-                .toList();
-    }
-}
