--- src/main/java/roomescape/reservationtime/dto/response/TimeInformation.java
@@ -1,9 +1,0 @@
-package roomescape.reservationtime.dto.response;
-
-import java.time.LocalTime;
-
-public record TimeInformation(
-        Long id,
-        LocalTime time
-) {
-}
