--- src/main/java/roomescape/reservationtime/dto/response/TimeSlotStatus.java
@@ -1,6 +1,0 @@
-package roomescape.reservationtime.dto.response;
-
-public enum TimeSlotStatus {
-    RESERVABLE,
-    WAITABLE
-}
