--- src/main/java/roomescape/reservationtime/presentation/ManagerReservationTimeController.java
@@ -1,47 +1,0 @@
-package roomescape.reservationtime.presentation;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import roomescape.common.api.ApiResponse;
-import roomescape.reservationtime.application.ReservationTimeService;
-import roomescape.reservationtime.dto.request.ReservationTimeSaveRequest;
-import roomescape.reservationtime.dto.response.ReservationTimeFindResponse;
-import roomescape.reservationtime.dto.response.ReservationTimeSaveResponse;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/manager/times")
-@RequiredArgsConstructor
-public class ManagerReservationTimeController {
-    private final ReservationTimeService reservationTimeService;
-
-    @PostMapping
-    public ResponseEntity<ApiResponse<ReservationTimeSaveResponse>> save(
-            @RequestBody @Valid ReservationTimeSaveRequest body
-    ) {
-        ReservationTimeSaveResponse response = reservationTimeService.save(body);
-        return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.success(response));
-    }
-
-    @GetMapping
-    public ResponseEntity<ApiResponse<List<ReservationTimeFindResponse>>> findAll() {
-        List<ReservationTimeFindResponse> responses = reservationTimeService.findAll();
-        return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success(responses));
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<Void>> delete(@PathVariable Long id) {
-        reservationTimeService.delete(id);
-        return ResponseEntity.noContent().build();
-    }
-}
