+++ src/main/java/roomescape/slot/application/SlotAssembler.java
@@ -7,12 +7,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
-import roomescape.reservationtime.ReservationTime;
-import roomescape.reservationtime.infrastructure.ReservationTimeRepository;
+import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.reservationtime.application.port.out.ReservationTimeRepository;
 import roomescape.slot.domain.Slot;
 import roomescape.slot.application.port.out.SlotRepository;
-import roomescape.theme.Theme;
-import roomescape.theme.infrastructure.ThemeRepository;
+import roomescape.theme.domain.Theme;
+import roomescape.theme.application.port.out.ThemeRepository;
 
 @Component
 @RequiredArgsConstructor
