--- src/main/java/roomescape/theme/Theme.java
@@ -1,13 +1,0 @@
-package roomescape.theme;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Theme {
-    private final Long id;
-    private String name;
-    private String description;
-    private String thumbnailUrl;
-}
