--- src/main/java/roomescape/theme/dto/request/ThemeSaveRequest.java
@@ -1,8 +1,0 @@
-package roomescape.theme.dto.request;
-
-public record ThemeSaveRequest(
-        String name,
-        String description,
-        String thumbnailUrl
-) {
-}
