--- src/main/java/roomescape/theme/dto/response/ThemeSaveResponse.java
@@ -1,19 +1,0 @@
-package roomescape.theme.dto.response;
-
-import roomescape.theme.Theme;
-
-public record ThemeSaveResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnailUrl
-) {
-    public static ThemeSaveResponse from(Theme theme) {
-        return new ThemeSaveResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnailUrl()
-        );
-    }
-}
