+++ src/test/java/roomescape/reservation/ReservationServiceTest.java
@@ -28,10 +28,10 @@ import roomescape.exception.EscapeRoomException;
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.application.dto.response.ReservationDetailFindResponse;
 import roomescape.reservation.application.port.out.ReservationRepository;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.slot.domain.Slot;
 import roomescape.slot.application.SlotAssembler;
-import roomescape.theme.Theme;
+import roomescape.theme.domain.Theme;
 import roomescape.waiting.domain.Waiting;
 import roomescape.waiting.domain.WaitingPromotionPolicy;
 import roomescape.waiting.application.port.out.WaitingRepository;
