+++ src/test/java/roomescape/reservationtime/JdbcReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.reservationtime.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.reservationtime.infrastructure.JdbcReservationTimeRepository;
+import roomescape.reservationtime.adapter.out.persistence.JdbcReservationTimeRepository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
