+++ src/test/java/roomescape/reservationtime/ReservationTimeApiIntegrationTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservationtime;
+package roomescape.reservationtime.domain;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
