+++ src/test/java/roomescape/waiting/WaitingServiceTest.java
@@ -9,10 +9,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
 import roomescape.reservation.application.port.out.ReservationRepository;
-import roomescape.reservationtime.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.slot.domain.Slot;
 import roomescape.slot.application.SlotAssembler;
-import roomescape.theme.Theme;
+import roomescape.theme.domain.Theme;
 import roomescape.waiting.application.WaitingService;
 import roomescape.waiting.application.dto.request.WaitingRequest;
 import roomescape.waiting.application.dto.response.WaitingResponse;
